// frontend/src/app/api/model/team-dto.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TeamDto { 
    id?: string;
    seasonId?: string;
    seasonName?: string;
    name?: string;
    description?: string;
    createdAt?: string;
    updatedAt?: string;
}